# openapi/qava.openapi.json
{
  "components": {
    "schemas": {
...
      "AuthoringDecisionInput": {
        "properties": {
          "blocking": {
            "default": true,
            "title": "Blocking",
            "type": "boolean"
          },
          "candidates": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Candidates",
            "type": "array"
          },
          "decision_id": {
            "minLength": 1,
            "title": "Decision Id",
            "type": "string"
          },
          "kind": {
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "resolution": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "status": {
            "default": "pending",
            "enum": [
              "pending",
              "confirmed",
              "overridden",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "decision_id",
          "kind"
        ],
        "title": "AuthoringDecisionInput",
        "type": "object"
      },
...
      "CreateDraftRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "output_contract": {
            "additionalProperties": true,
            "title": "Output Contract",
            "type": "object"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "output_contract"
        ],
        "title": "CreateDraftRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RawDraftRequest": {
        "properties": {
          "assistance_policy": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assistance Policy"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/AuthoringDecisionInput"
            },
            "title": "Decisions",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "health_policy": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Health Policy"
          },
          "output_contract": {
            "additionalProperties": true,
            "title": "Output Contract",
            "type": "object"
          },
          "questions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Questions",
            "type": "array"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "output_contract"
        ],
        "title": "RawDraftRequest",
        "type": "object"
      },
      "ResolveDecisionRequest": {
        "properties": {
          "action": {
            "enum": [
              "confirm",
              "override",
              "reject"
            ],
            "title": "Action",
            "type": "string"
          },
          "decision_id": {
            "minLength": 1,
            "title": "Decision Id",
            "type": "string"
          },
          "resolution": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          }
        },
        "required": [
          "decision_id",
          "action"
        ],
        "title": "ResolveDecisionRequest",
        "type": "object"
      },
...
      "UpdateDraftRequest": {
        "properties": {
          "operations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Operations",
            "type": "array"
          }
        },
        "required": [
          "operations"
        ],
        "title": "UpdateDraftRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Contract-first authoring, interviewing, projection, health, and explicit publication.",
    "title": "Qava MVP API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/questionnaire-drafts": {
      "post": {
        "operationId": "createQuestionnaireDraft",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Draft",
        "tags": [
          "questionnaires"
        ]
      }
    },
    "/api/v1/questionnaire-drafts/{draft_id}": {
      "delete": {
        "operationId": "deleteQuestionnaireDraft",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Draft",
        "tags": [
          "questionnaires"
        ]
      },
      "get": {
        "operationId": "getQuestionnaireDraft",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Draft",
        "tags": [
          "questionnaires"
        ]
      },
      "patch": {
        "operationId": "updateQuestionnaireDraft",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Draft",
        "tags": [
          "questionnaires"
        ]
      }
    },
    "/api/v1/questionnaire-drafts/{draft_id}/decisions": {
      "post": {
        "operationId": "resolveAuthoringDecision",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Decision",
        "tags": [
          "questionnaires"
        ]
      }
    },
    "/api/v1/questionnaire-drafts/{draft_id}/publish": {
      "post": {
        "operationId": "publishQuestionnaireDraft",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Draft",
        "tags": [
          "questionnaires"
        ]
      }
    },
    "/api/v1/questionnaire-drafts/{draft_id}/raw": {
      "put": {
        "operationId": "replaceQuestionnaireDraftRaw",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "title": "Draft Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RawDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Raw Draft",
        "tags": [
          "questionnaires"
        ]
      }
    },
    "/api/v1/questionnaires/{questionnaire_id}/versions/{version}": {
      "get": {
        "operationId": "get_published_questionnaire_api_v1_questionnaires__questionnaire_id__versions__version__get",
        "parameters": [
          {
            "in": "path",
            "name": "questionnaire_id",
            "required": true,
            "schema": {
              "title": "Questionnaire Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Published Questionnaire",
        "tags": [
          "questionnaires"
        ]
      }
    },
...
  }
}